Removed States (Is now redundant) Removed Namespaces Fixed Formatting Issues from Github's editor Added explict enum defs and index error checks
