+++ Digital 1/Assets/Scripts/PlayArea/PlayArea.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class PlayArea : MonoBehaviour
 {
 	/// <summary>
-	/// TEMPERARY variable for debugging purposes
+	/// TEMPORARY variable for debugging purposes
 	/// Can also be kept to ensure there is a background
 	/// </summary>
 	[SerializeField]
@@ -27,12 +27,12 @@ public class PlayArea : MonoBehaviour
 			return rect;
 		}
 	}
-  
-  /// <summary>
+
+	/// <summary>
 	/// 
 	/// </summary>
-  /// <remarks>Uses LateStart to fix cam not being correctly set.</remarks>
-  void LateStart()
+	/// <remarks>Uses LateStart to fix cam not being correctly set.</remarks>
+	void LateStart()
 	{
 		camera = Camera.main;
