+++ Digital 1/Assets/Scripts/Shared/ColoredObj.cs.meta
