--- Digital 1/Assets/Scripts/Shared/States.cs
@@ -1,10 +1,0 @@
-﻿/// <summary>
-/// List of potential color states of the player
-/// </summary>
-public enum ColorState
-{
-    Neutral,
-    Red,
-    Blue,
-    Yellow
-}
