--- Digital 1/Assets/Scripts/Shared/States.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0d0cf5b4892750e4f8492ec891e21f74
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
